+++ pdoc/web.py
@@ -12,7 +12,7 @@ import http.server
 import traceback
 import warnings
 import webbrowser
-from typing import Collection, Optional, Union
+from typing import Iterable, Iterator, Mapping, Optional, Union
 
 from pdoc import doc, extract, render
 from pdoc._compat import cache, removesuffix
@@ -44,19 +44,19 @@ class DocHandler(http.server.BaseHTTPRequestHandler):
             out = render.html_index(self.server.all_modules)
         elif path == "/search.js":
             self.send_response(200)
-            self.send_header("content-type", "application/json")
+            self.send_header("content-type", "application/javascript")
             self.end_headers()
             return self.server.render_search_index()
         else:
-            module = removesuffix(path.lstrip("/"), ".html").replace("/", ".")
-            if module not in self.server.all_modules:
+            module_name = removesuffix(path.lstrip("/"), ".html").replace("/", ".")
+            if module_name not in self.server.all_modules:
                 self.send_response(404)
                 self.send_header("content-type", "text/html")
                 self.end_headers()
-                return render.html_error(error=f"Module {module!r} not found")
+                return render.html_error(error=f"Module {module_name!r} not found")
 
             mtime = ""
-            t = extract.module_mtime(module)
+            t = extract.module_mtime(module_name)
             if t:
                 mtime = f"{t:.1f}"
             if "mtime=1" in self.path:
@@ -66,8 +66,8 @@ class DocHandler(http.server.BaseHTTPRequestHandler):
                 return mtime
 
             try:
-                extract.invalidate_caches(module)
-                mod = doc.Module(extract.load_module(module))
+                extract.invalidate_caches(module_name)
+                mod = self.server.all_modules[module_name]
                 out = render.html_module(
                     module=mod,
                     all_modules=self.server.all_modules,
@@ -78,7 +78,7 @@ class DocHandler(http.server.BaseHTTPRequestHandler):
                 self.send_header("content-type", "text/html")
                 self.end_headers()
                 return render.html_error(
-                    error=f"Error importing {module!r}",
+                    error=f"Error importing {module_name!r}",
                     details=traceback.format_exc(),
                 )
 
@@ -97,24 +97,54 @@ class DocHandler(http.server.BaseHTTPRequestHandler):
 class DocServer(http.server.HTTPServer):
     """pdoc's live-reloading web server"""
 
-    all_modules: Collection[str]
+    all_modules: AllModules
 
-    def __init__(self, addr: tuple[str, int], all_modules: Collection[str], **kwargs):
+    def __init__(self, addr: tuple[str, int], specs: list[str], **kwargs):
         super().__init__(addr, DocHandler, **kwargs)  # type: ignore
-        self.all_modules = all_modules
+        module_names = extract.walk_specs(specs)
+        self.all_modules = AllModules(module_names)
 
     @cache
     def render_search_index(self) -> str:
         """Render the search index. For performance reasons this is always cached."""
-        all_mods = {}
+        # Some modules may not be importable, which means that they would raise an RuntimeError
+        # when accessed. We "fix" this by pre-loading all modules here and only passing the ones that work.
+        all_modules_safe = {}
         for mod in self.all_modules:
             try:
-                m = extract.load_module(mod)
+                all_modules_safe[mod] = doc.Module.from_name(mod)
             except RuntimeError:
                 warnings.warn(f"Error importing {mod!r}:\n{traceback.format_exc()}")
-            else:
-                all_mods[mod] = doc.Module(m)
-        return render.search_index(all_mods)
+        return render.search_index(all_modules_safe)
+
+
+class AllModules(Mapping[str, doc.Module]):
+    """A lazy-loading implementation of all_modules.
+
+    This behaves like a regular dict, but modules are only imported on demand for performance reasons.
+    This has the somewhat annoying side-effect that __getitem__ may raise a RuntimeError.
+    We can ignore that when rendering HTML as the default templates do not access all_modules values,
+    but we need to perform additional steps for the search index.
+    """
+
+    def __init__(self, allowed_modules: Iterable[str]):
+        # use a dict to preserve order
+        self.allowed_modules: dict[str, None] = dict.fromkeys(allowed_modules)
+
+    def __len__(self) -> int:
+        return self.allowed_modules.__len__()
+
+    def __iter__(self) -> Iterator[str]:
+        return self.allowed_modules.__iter__()
+
+    def __contains__(self, item):
+        return self.allowed_modules.__contains__(item)
+
+    def __getitem__(self, item: str):
+        if item in self.allowed_modules:
+            return doc.Module.from_name(item)
+        else:  # pragma: no cover
+            raise KeyError(item)
 
 
 # https://github.com/mitmproxy/mitmproxy/blob/af3dfac85541ce06c0e3302a4ba495fe3c77b18a/mitmproxy/tools/web/webaddons.py#L35-L61