+++ test/testdata/demopackage/child_excluded.py
@@ -1,0 +1,4 @@
+"""child_excluded docstr
+
+This submodule is excluded via specs and shouldn't show up in the submodules list.
+"""